requirements.txt
streamlit
ultralytics
torch
opencv-python
pillow
numpy